requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
Pillow==10.0.0
python-decouple==3.8
djoser==2.2.0
django-cors-headers==4.2.0
factory-boy==3.3.0
django-filter==23.2
gunicorn==21.2.0
toml==0.10.2
black==23.9.1
whitenoise==6.6.0
